refactor: normalize tags to lib
