--- app/lib/filters.ts
@@ -1,3 +1,0 @@
-export function normalizeTags(tags: string | string[]): string[] {
-  return Array.isArray(tags) ? tags : [tags];
-}
