+++ lib/utils.ts
@@ -4,3 +4,6 @@ import { twMerge } from "tailwind-merge";
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
+export function normalizeTags(tags: string | string[]): string[] {
+  return Array.isArray(tags) ? tags : [tags];
+}
